style: resolve SwiftLint violations and refactor tests
