+++ Sources/tyme/lunar/LunarDay.swift
@@ -13,7 +13,11 @@ import Foundation
 /// let solarDay = day.solarDay                               // Convert to solar
 /// ```
 public final class LunarDay: DayUnit, Tyme {
-    public static let NAMES = ["初一", "初二", "初三", "初四", "初五", "初六", "初七", "初八", "初九", "初十", "十一", "十二", "十三", "十四", "十五", "十六", "十七", "十八", "十九", "二十", "廿一", "廿二", "廿三", "廿四", "廿五", "廿六", "廿七", "廿八", "廿九", "三十"]
+    public static let NAMES = [
+        "初一", "初二", "初三", "初四", "初五", "初六", "初七", "初八", "初九", "初十",
+        "十一", "十二", "十三", "十四", "十五", "十六", "十七", "十八", "十九", "二十",
+        "廿一", "廿二", "廿三", "廿四", "廿五", "廿六", "廿七", "廿八", "廿九", "三十"
+    ]
 
     /// Whether this day falls in a leap month (闰月 Rùnyuè).
     public let leap: Bool
